+++ lib/multitasking.dart
@@ -1,4 +1,5 @@
 export 'src/multitasking/cancellation.dart';
 export 'src/multitasking/errors.dart';
+export 'src/multitasking/isolate_runner.dart';
 export 'src/multitasking/stream.dart';
 export 'src/multitasking/task.dart';